UHF-9690: Use CharacterCounterFieldWidgetTrait.
